# Conflicts:
#	comcer_app/lib/dominio/enum/order_status.dart
#	comcer_app/lib/util/constants.dart
#	comcer_app/lib/views/components/card/order_in_progress_card/OrderCard.dart
#	comcer_app/lib/views/components/card/order_product_card.dart
#	comcer_app/lib/views/home_screen.dart
#	comcer_app/lib/views/login_screen.dart
#	comcer_app/lib/views/orders_screen.dart
#	comcer_app/lib/views/products_of_order_screen.dart

+++ comcer_app/lib/dominio/models/order_pad.dart
@@ -1,6 +1,6 @@
-import 'package:comcer_app/dominio/models/BaseAPIResponse.dart';
+import 'package:comcer_app/dominio/models/base_api_response.dart';
 
-import 'inconsistencia_validacao.dart';
+import 'inconsistencies.dart';
 import 'order.dart';
 
 class OrderPad extends BaseAPIResponse {
@@ -64,7 +64,7 @@ class OrderPad extends BaseAPIResponse {
         _listaPedidos.add(Order.fromJson(order));
       });
     }
-    _valor = json['valor'];
+    _valor = json['valor'].toDouble();
     _status = json['status'];
   }
